populated gold
created a vector of pairs to keep track of invalid coordinates.
Added generateRandomLocation() and invalidCoord() - they don't work for populateBoulder for now
Gold might be populated in middle - couldn't find if that's allowed in specs.
To see the populated gold uncomment setVisible(true) in Actor.h and set depth to 0 in Acquirables in Actor.h