iceman can remove ice. BUG in description
When iceman moves along the row in which he spawns, ice is removed from the bottom. I believe this has something to do with the fact that the map and the 2D array y-coordinates are inverted. I noticed that the coordinates icePtr[30][0] are the same as icePtr[29][60]. Very weird. [29][60] should be invalid since the ice is only 59 units high, but it instead accesses the 0'th row. The x coordinate is also ahead by 1.